pyproject.toml
[project]
name = "avalokiteshvara"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "httpx>=0.27",
    "matplotlib>=3.10.8",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=20.0",
    "requests>=2.32.5",
]
